refactor(pipeline): move process_line to processors module
Keep the script as a thin CLI wrapper — process_line is pipeline logic,
not orchestration.
